+++ tests/xua_unit_tests/src/xua_unit_tests.h
@@ -1,0 +1,9 @@
+// Copyright 2021 XMOS LIMITED.
+// This Software is subject to the terms of the XMOS Public Licence: Version 1.
+#ifndef XUA_UNIT_TESTS_H_
+#define XUA_UNIT_TESTS_H_
+
+#include "unity.h"
+#include "xua_conf.h"
+
+#endif /* XUA_UNIT_TESTS_H_ */